requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
tqdm==4.62.3
matplotlib==3.5.1
numpy>=1.19.5
pillow>=10.2.0
opencv-python>=4.5.1.48
torch==1.8.0+cu111
torchvision==0.9.0+cu111
torchaudio==0.8.0
scipy==1.7.3
requests==2.26.0
termcolor==1.1.0
easydict==1.9.0
seaborn==0.11.2
imageio==2.16.0
lpips==0.1.4
scikit-learn==1.3.2 
